requirements_clean.txt
# Core dependencies for Adaptive Photo Assistant
# Computer Vision and Image Processing
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
pillow>=10.0.0

# Voice and Speech Processing
pyttsx3>=2.90
SpeechRecognition>=3.10.0
pyaudio>=0.2.11

# AI and API Communication
requests>=2.31.0
anthropic>=0.5.0

# Web Interface (Optional)
flask>=2.3.0
flask-cors>=4.0.0

# Utility Libraries
python-dateutil>=2.8.0
packaging>=23.0
python-dotenv>=1.0.0

# Development and Testing (Optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Note: Some packages may require system dependencies:
# - PyAudio may need portaudio development headers
# - MediaPipe requires specific system libraries
# - For Windows: Consider using conda or pre-built wheels
